# swagger/swagger.yaml
basePath: /v1/
consumes:
- application/json
definitions:
  AodAPIModel:
    properties:
      date:
        description: The date when this author was the author of the day
        example: "2021-06-12T00:00:00Z"
        type: string
        x-go-name: Date
      id:
        description: The author's id
        example: 24952
        format: int64
        type: integer
        x-go-name: Id
      name:
        description: The name of the author
        example: Muhammad Ali
        type: string
        x-go-name: Name
    type: object
    x-go-package: _/Users/thorduragustsson/Programming/go/github.com/Skjaldbaka17/quotel-sls-api/swagger/docs
  AuthorAPIModel:
    properties:
      count:
        description: The popularity index of the author
        example: 1111
        format: int64
        type: integer
        x-go-name: Count
      hasIcelandicQuotes:
        description: Whether or not this author has some icelandic quotes
        example: true
        type: boolean
        x-go-name: HasIcelandicQuotes
      id:
        description: The author's id
        example: 24952
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: Id
      name:
        description: Name of the author
        example: Muhammad Ali
        type: string
        x-go-name: Name
      nrOfEnglishQuotes:
        description: How many quotes in English this author has
        example: 78
        format: int64
        type: integer
        x-go-name: NrOfEnglishQuotes
      nrOfIcelandicQuotes:
        description: How many quotes in Icelandic this author has
        example: 6
        format: int64
        type: integer
        x-go-name: NrOfIcelandicQuotes
    type: object
    x-go-package: _/Users/thorduragustsson/Programming/go/github.com/Skjaldbaka17/quotel-sls-api/swagger/docs
  OfTheDayModel:
    properties:
      date:
        description: The date when this author / quote should be 'of the day'
        example: "2020-06-12"
        type: string
        x-go-name: Date
      id:
        description: The id of the author / quote
        example: 1
        format: int64
        type: integer
        x-go-name: Id
      language:
        default: English
        description: The language of this author / quote
        example: icelandic
        type: string
        x-go-name: Language
    type: object
    x-go-name: ofTheDayModel
    x-go-package: _/Users/thorduragustsson/Programming/go/github.com/Skjaldbaka17/quotel-sls-api/swagger/docs
...
  QodViewAPIModel:
    properties:
      authorId:
        description: The author's id
        format: int64
        type: integer
        x-go-name: AuthorId
      date:
        description: The date when this quote was the quote of the day
        example: "2021-06-12T00:00:00Z"
        type: string
        x-go-name: Date
      isIcelandic:
        description: |-
          Whether the quote is in icelandic
          false
        type: boolean
        x-go-name: IsIcelandic
      name:
        description: Name of the author
        example: Muhammad Ali
        type: string
        x-go-name: Name
      quote:
        description: The quote for the day
        example: Float like a butterfly, sting like a bee
        type: string
        x-go-name: Quote
      quoteId:
        description: The quote's id
        example: 582676
        format: int64
        type: integer
        x-go-name: QuoteId
    type: object
    x-go-package: _/Users/thorduragustsson/Programming/go/github.com/Skjaldbaka17/quotel-sls-api/swagger/docs
  SearchViewAPIModel:
    properties:
      authorId:
        description: The author's id
        example: 24952
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: AuthorId
      isIcelandic:
        description: Whether or not this quote is in Icelandic or not
        example: false
        type: boolean
        x-go-name: IsIcelandic
      name:
        description: Name of author
        example: Muhammad Ali
        type: string
        x-go-name: Name
      quote:
        description: The quote
        example: Float like a butterfly, sting like a bee.
        type: string
        x-go-name: Quote
      quoteId:
        description: The quote's id
        example: 582676
        format: int64
        type: integer
        uniqueItems: true
        x-go-name: QuoteId
    type: object
    x-go-package: _/Users/thorduragustsson/Programming/go/github.com/Skjaldbaka17/quotel-sls-api/swagger/docs
...
host: rxvglshzhl.execute-api.eu-west-1.amazonaws.com
info:
  contact:
    email: skjaldbaka17@gmail.com
    name: Þórður Ágústsson
  description: Quotel API offers a complete feature rich REST API access to all your
    quotes needs. This is the documentation of our quotes API.
  title: quotes-api.
  version: 1.0.0
paths:
  /authors:
    post:
      description: Get the authors by their ids
      operationId: GetAuthors
      parameters:
      - description: The structure of the request for getting authors by their ids
        in: body
        name: Body
        required: true
        schema:
          properties:
            ids:
              description: A list of the authors's ids that you want
              example:
              - 24952
              - 19161
              items:
                format: int64
                type: integer
              type: array
              x-go-name: Ids
          required:
          - ids
          type: object
      responses:
        "200":
          $ref: '#/responses/authorsResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - AUTHORS
  /authors/aod:
    post:
      description: Gets the author of the day
      operationId: GetAuthorOfTheDay
      parameters:
      - description: The structure of the request for getting the author / quote of
          the day
        in: body
        name: Body
        schema:
          properties:
            language:
              default: English
              description: Get the author / quote of the day for the given language
                ("icelandic" or "english")
              example: English
              type: string
              x-go-name: Language
          type: object
      responses:
        "200":
          $ref: '#/responses/aodResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - AUTHORS
  /authors/aod/history:
    post:
      description: Gets the history for the authors of the day
      operationId: GetAODHistory
      parameters:
      - description: The structure of the request for getting the history of AODs
          / QODs
        in: body
        name: Body
        schema:
          items:
            properties:
              language:
                default: English
                description: Get the history of the AODS / QODs for the given language
                  ("icelandic" or "english")
                example: icelandic
                type: string
                x-go-name: Language
              minimum:
                description: The earliest date to return. All authors / quotes between
                  minimum and today will be returned.
                example: "2020-12-21"
                type: string
                x-go-name: Minimum
            type: object
          type: array
      responses:
        "200":
          $ref: '#/responses/aodHistoryResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - AUTHORS
  /authors/aod/new:
    post:
      description: sets the author of the day for the given dates
      operationId: SetAuthorOfTheDay
      parameters:
      - description: The structure of the request for setting AODs
        in: body
        name: Body
        schema:
          items:
            properties:
              aods:
                items:
                  $ref: '#/definitions/OfTheDayModel'
                type: array
                x-go-name: Aods
            type: object
          type: array
      responses:
        "200":
          $ref: '#/responses/successResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - AUTHORS
...
  /authors/random:
    post:
      description: Get a random Author, and some of his quotes, according to the given
        parameters
      operationId: GetRandomAuthor
      parameters:
      - description: The structure of the request for getting a random author
        in: body
        name: Body
        schema:
          properties:
            language:
              description: |-
                The random author must have quotes in the given language ("english" or "icelandic") if left empty then no
                constraint on language is set
              example: English
              type: string
              x-go-name: Language
            maxQuotes:
              default: 1
              description: How many quotes, maximum, to be returned from this author
              example: 10
              format: int64
              maximum: 50
              type: integer
              x-go-name: MaxQuotes
          type: object
      responses:
        "200":
          $ref: '#/responses/searchViewsResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - AUTHORS
...
  /quotes/qod/history:
    post:
      description: Gets the history for the quotes of the day
      operationId: GetQODHistory
      parameters:
      - description: The structure of the request for getting the history of AODs
          / QODs
        in: body
        name: Body
        schema:
          items:
            properties:
              language:
                default: English
                description: Get the history of the AODS / QODs for the given language
                  ("icelandic" or "english")
                example: icelandic
                type: string
                x-go-name: Language
              minimum:
                description: The earliest date to return. All authors / quotes between
                  minimum and today will be returned.
                example: "2020-12-21"
                type: string
                x-go-name: Minimum
            type: object
          type: array
      responses:
        "200":
          $ref: '#/responses/qodHistoryResponse'
        "400":
          $ref: '#/responses/incorrectBodyStructureResponse'
        "500":
          $ref: '#/responses/internalServerErrorResponse'
      tags:
      - QUOTES
...
produces:
- application/json
responses:
  aodHistoryResponse:
    description: Data structure representing the response for the history of AODs
    schema:
      items:
        $ref: '#/definitions/AodAPIModel'
      type: array
  aodResponse:
    description: Data structure representing the response for the author of the day
    schema:
      $ref: '#/definitions/AodAPIModel'
  authorsResponse:
    description: Data structure representing the response for authors
    schema:
      items:
        $ref: '#/definitions/AuthorAPIModel'
      type: array
  incorrectBodyStructureResponse:
    description: Data structure representing the error response to a wrongly structured
      request body
    schema:
      properties:
        message:
          description: The error message
          example: request body is not structured correctly.
          type: string
          x-go-name: Message
      type: object
...
  internalServerErrorResponse:
    description: Data structure representing the error response to an internal server
      error
    schema:
      properties:
        message:
          description: The error message
          example: Please try again later.
          type: string
          x-go-name: Message
      type: object
...
  qodHistoryResponse:
    description: Data structure representing the response for the history of QODS
    schema:
      items:
        $ref: '#/definitions/QodViewAPIModel'
      type: array
...
  searchViewsResponse:
    description: Data structure representing the response for quotes
    schema:
      items:
        $ref: '#/definitions/SearchViewAPIModel'
      type: array
  successResponse:
    description: ""
    schema:
      properties:
        message:
          example: This request was a success
          type: string
          x-go-name: Message
        statusCode:
          description: HTTP status code
          example: 200
          format: int64
          type: integer
          x-go-name: StatusCode
      type: object
...
schemes:
- https
swagger: "2.0"
...